Created a database in Mysql for the app, and I manually created the relevant tables. For now it is only the members table. I also installed "MySql.Data" in Visual Studio.
I also Created a new class called DBConnection to connect to the database so if the other classes in the app want access to the database then they can only call an object of that class.

I also tested the connection in the Form1 and it actually worked. Lastly I added comments in the DBConnection Class for clarity.